SUbmitting mandates by user name works.
No need to enter ID, just input the name and the system will find the appropriate user. We can copy the same system for roles, and we also need to add dropdowns that are sent along.

ALso, any spelling mistake will result in failure, so we need to add along a dropdown (or select) to minimize user error. We need to:

 * Add dropdown/select
 * Handle failure
 * Use same system for roles
 * See if we can compare case-insensitively
 * Think about name clashes?