Fix duplicated middle names
The problem seems to have been that we did get `givenName` from hodis,
which contains both first names and middle names, but we do not get `sn`
(which doesn't exist in hodis, but maybe it once existed in zfinger?),
so instead we tried to extract the last name from `cn` by just removing
what was before the first space (so it did remove middle names).

This fix isn't perfect, but should work better than before.

Also note that this will remove the (kthid) at the end of peoples'
last names.